upload file when creating, update docs
